refactor: make refactoring, add space functional
