+++ keyboard/src/index.js
@@ -1,14 +1,26 @@
 
-import {  renderTextAreaToDom, renderKeyBoardToDom, animationKeyboardClick,  handleClick, createLangChangeCommandBtn, activeBackspace, keyboardClick} from "./js/keys.js";
+import {
+  language,
+  shift,
+  renderTextAreaToDom,
+  renderRowsCollectionToDom,
+  renderKeyToDom,
+  animationKeyboardClick,
+  keyboardClick,
+} from "./js/keys-copy.js";
+
 window.onload = function () {
-  renderTextAreaToDom()
-  renderKeyBoardToDom()
+
+  language,
+  shift,
+
+  renderTextAreaToDom(),
+  renderRowsCollectionToDom(),
+  renderKeyToDom(language, shift),
   animationKeyboardClick()
-  handleClick()
-  createLangChangeCommandBtn()
-  activeBackspace()
   keyboardClick()
 }
+
