--- keyboard/src/js/KeyRow.js
@@ -1,16 +1,0 @@
-export class KeyRow {
-  constructor(className) {
-    this.className = className;
-  }
-
-  // KeyRow generator
-  generateKeyRow() {
-    let template = "";
-    let rowElement = document.createElement("div");
-    rowElement.className = `row ${this.className}`;
-    rowElement.innerHTML = template;
-    return rowElement;
-  }
-}
-
-
